Move file constants to backend code
